Breakfast: Bake: Introduce v1.0
'bake.sh' - a simple shell script, which greatly simplifies the compilation and assembly of the Linux kernel. The main goal of 'bake', a part of the project 'Breakfast', - to automatize the process of creating a kernel.

The current version of bake - 1.0 - gives the opportunity to create virtually any kernel for mobile devices. Of course, this script still has potential. For example - a more profound adaptation to the 64-bit architecture, in particular - AARCH64 (ARMv8.x family), namely - a fully automatic compiling of the 'dt' image. However, I have the time and desire to maintain the breakfast, so all this stuff will be in the future.

The script logic is simple and obvious. For a better understanding of the processes I've added a lot of comments (in order to avoid confusion). Lines that the developer should change are allocated to the group 'Userspace configuration'. In general, I hope there will be no problems.

The script was tested on ARM (Linux 3.4 hammerhead), where it showed a stable and proper work. If there are any bugs / suggestions / etc - please let me know by email.

Signed-off-by: Alex Saiko <solcmdr@gmail.com>