Bake: Improve output file generation
I've noticed the inconsistence in date setting after a month switch. This happens due to
the wrong date declaration, which uses DayMonth as the first symbols. Fix it by properly generate
current date in YearMonthDay format.

One of the main features of Breakfast system is its parallel building. But this concept ruins
because of the united output directory. Collect built stuff in its own directory.

- Reverse date selection,
- Export output files to their SOURCE-dependent directories,
- Correct code comments;

Signed-off-by: Alex Saiko <solcmdr@gmail.com>