Breakfast: Bake: Introduce v3.0.0
This is a huge rewrite of the whole bake script.

First of all, I got rid of source file, which was pretty useless
considering the fact that nothing except bake.sh itself was using
it. All the functions were written from scratch (except cleaning
one) in a better language.

Bake's functionality is significantly reduces, but its main feature
(multitarget support) is here. You should run bake via
 "$ ./bake.sh <configuration_name>" or "$ sh bake.sh <configuration_name>"
where "<configuration_name" is a full or a part name of the Bake
configuration file. An input configuration file will be hooked by
find_config function, which uses "find" command to determine the
selected config's full name. Because of this, you can use shortcuts
like "h" instead of "hammerhead.conf". Note that find command
is used with a flag "-quit" meaning it will return the path (name
in our case) for the first match only. If you have multiple configs
with the same start (like "hammerhead.conf" and "hammerheadcaf.conf")
the config which is lower in alphabetical order will be hooked.

Configuration files itself are much simpler now because useless
variables were removed and printing was moved to the main script.

The only function of the script is building and packaging the
kernel for a specific target. More functions (and previous
functionality) will become later.

Change-Id: I2003f04b313d10bbb7ed6562fbc0654a2125f2ad
Signed-off-by: Alex Saiko <solcmdr@gmail.com>