Breakfast: Bake: Improve make_flashable
While testing the script by using it as a source I have noticed
that make_flashable moves to a root (/) directory after processing.
This happens because of the typo error at the end of the function.

Fix it by correcting the typo in cd command.

Also add a protecting in case that we don't have  file
in its place. This will fix an issue when we don't have the kernel
image itself but make_flashable continues to process.

Change-Id: I8a651f31eefeafac2b833466a9227e4567831fbe
Signed-off-by: Alex Saiko <solcmdr@gmail.com>