Breakfast: Bake: Introduce v3.2
This is a cleanup of unused code with some miscellaneous adjustements and
freshly added useless commentaries. Also I've reduced the duplication of
flashable installer code by unifying it between all supported targets. Now
the installer code is completely device-independent (it was to be so
previously too, by the way) and is placed in a separate directory called
breakfast-installer.  Device-dependent scripts and resources will be
squashed with installer during the process of flashable archive creation,
hence the functionality is untouched.

Change-Id: Ifab123f9e0fef5b4044939c2039339695b37645a
Signed-off-by: Alex Saiko <solcmdr@gmail.com>