Breakfast: Bake: Introduce v3.3
After some time I became confused with the names of some internal
variables because they are not imformative.  A while ago the eyes
treated this fact not so critical due to a freshment of a new naming,
but now it is obvious that something must be done with it.

So, I have rewritten all the internal variable names to get rid of this
5-char limitation as it is extremely uncosy to maintain.  Also I have
adjusted some pieces of code, including termination function that became
readable, config checker, which is able to handle void declarations now
(this is truth for every variable defined check, though).

Signing function has been renamed to sign_package() and is completely
exportable now, which means that it is finally possible to sign every
supported package by simply calling sign_package() with that package's
name as an argument. Much cosier now.

Some miscellaneous stuff was also made, as always.

Change-Id: I33fc4d0ab76ffccac8d6716b816b209683833b0d
Signed-off-by: Alex Saiko <solcmdr@gmail.com>